// src/app/data-access/api/api-service-base.service.ts
export class AuthResponse implements IAuthResponse {
  accessToken?: string | undefined;
  refreshToken?: string | undefined;
  expiresAt?: Date | undefined;
  user?: UserDto;
  remainingAttempts?: number | undefined;
  lockoutEnd?: Date | undefined;

  constructor(data?: IAuthResponse) {
    if (data) {
      for (var property in data) {
        if (data.hasOwnProperty(property))
          (this as any)[property] = (data as any)[property];
      }
    }
  }

  init(_data?: any) {
    if (_data) {
      this.accessToken = _data['accessToken'];
      this.refreshToken = _data['refreshToken'];
      this.expiresAt = _data['expiresAt']
        ? new Date(_data['expiresAt'].toString())
        : (undefined as any);
      this.user = _data['user']
        ? UserDto.fromJS(_data['user'])
        : (undefined as any);
      this.remainingAttempts = _data['remainingAttempts'];
      this.lockoutEnd = _data['lockoutEnd']
        ? new Date(_data['lockoutEnd'].toString())
        : (undefined as any);
    }
  }

  static fromJS(data: any): AuthResponse {
    data = typeof data === 'object' ? data : {};
    let result = new AuthResponse();
    result.init(data);
    return result;
  }

  toJSON(data?: any) {
    data = typeof data === 'object' ? data : {};
    data['accessToken'] = this.accessToken;
    data['refreshToken'] = this.refreshToken;
    data['expiresAt'] = this.expiresAt
      ? this.expiresAt.toISOString()
      : (undefined as any);
    data['user'] = this.user ? this.user.toJSON() : (undefined as any);
    data['remainingAttempts'] = this.remainingAttempts;
    data['lockoutEnd'] = this.lockoutEnd
      ? this.lockoutEnd.toISOString()
      : (undefined as any);
    return data;
  }
}

export interface IAuthResponse {
  accessToken?: string | undefined;
  refreshToken?: string | undefined;
  expiresAt?: Date | undefined;
  user?: UserDto;
  remainingAttempts?: number | undefined;
  lockoutEnd?: Date | undefined;
}
// ...
export class UserDto implements IUserDto {
  id?: string;
  email?: string | undefined;
  fullName?: string | undefined;
  roles?: string[] | undefined;

  constructor(data?: IUserDto) {
    if (data) {
      for (var property in data) {
        if (data.hasOwnProperty(property))
          (this as any)[property] = (data as any)[property];
      }
    }
  }

  init(_data?: any) {
    if (_data) {
      this.id = _data['id'];
      this.email = _data['email'];
      this.fullName = _data['fullName'];
      if (Array.isArray(_data['roles'])) {
        this.roles = [] as any;
        for (let item of _data['roles']) this.roles!.push(item);
      }
    }
  }

  static fromJS(data: any): UserDto {
    data = typeof data === 'object' ? data : {};
    let result = new UserDto();
    result.init(data);
    return result;
  }

  toJSON(data?: any) {
    data = typeof data === 'object' ? data : {};
    data['id'] = this.id;
    data['email'] = this.email;
    data['fullName'] = this.fullName;
    if (Array.isArray(this.roles)) {
      data['roles'] = [];
      for (let item of this.roles) data['roles'].push(item);
    }
    return data;
  }
}

export interface IUserDto {
  id?: string;
  email?: string | undefined;
  fullName?: string | undefined;
  roles?: string[] | undefined;
}